requirements.txt
langchain-core>=0.1.52
langchain-openai>=0.1.3
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
duckduckgo-search>=5.0.0